relocate UseOperation package name
